Device, Location

+++ MMSProject/src/main/java/com/husony/repository/DeviceRepository.java
@@ -14,5 +14,5 @@ import java.util.List;
 public interface DeviceRepository {
     List<Device> getDevices();
     void addOrUpdate(Device d);
-    Device getDeviceById(int id);
+    Device getDeviceById(long id);
 }
